ACP_V1/requirements_master.txt
chromadb
fastapi==0.104.0
fastapi>=0.110.0
fitz
flask
gunicorn
h5py
jax
jaxlib
jsonschema>=4.0.0
matplotlib
numpy
openai
opencv-python
pandas
pandas>=2.0.0
paramiko>=2.11.0
persim
pillow
pydantic==2.4.0
pydantic>=2
pytesseract
pytest-cov>=4.1.0
pytest>=7.4.0
python-dotenv
pyyaml
pyyaml>=6.0
requests>=2.31.0
ripser
scipy
sqlite-utils>=3.30
streamlit>=1.24.0
tiktoken
types-requests>=2.31.0
uvicorn==0.24.0
uvicorn[standard]>=0.24.0
watchdog
watchdog>=3.0.0